+++ 6_2_1/ex1/src/custom_int.cpp
@@ -1,4 +1,5 @@
 #include "../include/custom_int.hpp"
+#include <cstdint>
 #include <iostream>
 #include <stdexcept>
 
@@ -11,31 +12,31 @@ CustomInt::CustomInt() {
   ++counter;
 }
 // Operation overloads
-int &CustomInt::operator=(CustomInt const &obj) {
+int64_t &CustomInt::operator=(CustomInt const &obj) {
   this->m_value = obj.m_value;
   return this->m_value;
 }
-int &CustomInt::operator=(const int val) {
+int64_t &CustomInt::operator=(const int val) {
   this->m_value = val;
   return this->m_value;
 }
-int &CustomInt::operator+(CustomInt const &obj) {
+int64_t &CustomInt::operator+(CustomInt const &obj) {
   this->m_value += obj.m_value;
     
   return this->m_value;
 }
 
-int &CustomInt::operator+(const int val) { return this->m_value += val; }
-int &CustomInt::operator+=(CustomInt const &obj) {
+int64_t &CustomInt::operator+(const int val) { return this->m_value += val; }
+int64_t &CustomInt::operator+=(CustomInt const &obj) {
   this->m_value += obj.m_value;
   return this->m_value;
 }
 
-int &CustomInt::operator+=(const int val) {
+int64_t &CustomInt::operator+=(const int val) {
   this->m_value += val;
   return this->m_value;
 }
-int &CustomInt::operator-(CustomInt const &obj) {
+int64_t &CustomInt::operator-(CustomInt const &obj) {
    this->m_value -= obj.m_value;
    if(this->m_value < 0){
     this->m_value = 0;
@@ -43,21 +44,21 @@ int &CustomInt::operator-(CustomInt const &obj) {
    return this->m_value;
 }
 
-int &CustomInt::operator-(const int val) {
+int64_t &CustomInt::operator-(const int val) {
    this->m_value -= val;
    if(this->m_value < 0){
     this->m_value = 0;
    }
    return this->m_value;
 }
-int &CustomInt::operator-=(CustomInt const &obj) {
+int64_t &CustomInt::operator-=(CustomInt const &obj) {
    this->m_value -= obj.m_value;
    if(this->m_value < 0){
     this->m_value = 0;
    }
    return this->m_value;
 }
-int &CustomInt::operator-=(const int val) {
+int64_t &CustomInt::operator-=(const int val) {
    this->m_value -= val;
    if(this->m_value < 0){
     this->m_value = 0;
@@ -66,7 +67,7 @@ int &CustomInt::operator-=(const int val) {
 }
 // end Operation overloads
 void CustomInt::set_value(const int value) { m_value = value; }
-const int &CustomInt::get_value() const { return m_value; }
+const int64_t &CustomInt::get_value() const { return m_value; }
 void CustomInt::print() const { std::cout << m_value << '\n'; }
 void CustomInt::print_counter_value() const { std::cout << counter << '\n'; }
